+++ src/main/java/org/nette/latte/psi/elements/LatteLinkPartElement.java
@@ -1,16 +1,15 @@
 package org.nette.latte.psi.elements;
 
-import com.intellij.openapi.util.TextRange;
 import com.intellij.psi.PsiElement;
+import com.intellij.util.IncorrectOperationException;
 import org.jetbrains.annotations.NotNull;
 
-public interface LatteLinkPartElement extends PsiElement {
+public interface LatteLinkPartElement extends LattePsiNamedElement {
     @NotNull
     LatteLinkElement getParentLink();
 
-    @NotNull
-    String getPartText();
-
-    @NotNull
-    TextRange getRangeInParent();
+    @Override
+    default PsiElement setName(@NotNull String name) throws IncorrectOperationException {
+        return this;
+    }
 }